ISOs: update checksum to latest, pin version
per #61, using a more specific URL will allow people to install
a slightly-out-of-date version as long as it is hosted,
after which the error will be that the release no longer exists rather
than that its checksum has failed.

This has a side benefit of happening failing more quickly,
before the user has downloaded a full disk image.